Fixed error with new data
New data had weighted and unweighted categories, which would double count the data. This actually produced an error in weeks not having matching sizes since the data for the last week in December was not in yet. The program should run smoothly with this and new data given that the format does not change drastically.